requirements.txt
Flask==3.0.0
flask-cors==4.0.0
chromadb==0.4.15
openai==1.35.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
any-agent==0.1.0
smolagents==0.1.0
gunicorn==21.2.0